apiServer/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pandas
fpdf2